Signs type fixed

+++ Plugins/NativeRobotics/Utils/Editor/RobotVisualizerGeneratorWindow.cs
@@ -52,7 +52,7 @@ namespace NativeRobotics.Utils.Editor
                 // Adding Mesh Robot View component
                 var view = robotMeshInstance.AddComponent<MeshRobotView>();
 
-                var signs = new float[joints.Count];
+                var signs = new int[joints.Count];
                 for (var i = 0; i < joints.Count; i++)
                 {
                     signs[i] = 1;
